envoy: move EnvoyServiceBackendSyncer from watchers to envoy package
After removing the dependency from the service package (service manager)
to envoy, it's possible to move the EnvoyServiceBackendSyncer from the
k8s watchers package to the envoy package.
